move walletIds file to constants folder
