+++ packages/wallets/src/evm/wallets/coinbase-wallet.ts
@@ -1,6 +1,6 @@
 import type { CoinbaseWalletConnector } from "../connectors/coinbase-wallet";
 import { TWConnector, WagmiAdapter } from "../interfaces/tw-connector";
-import { walletIds } from "../walletIds";
+import { walletIds } from "../constants/walletIds";
 import { AbstractClientWallet, WalletOptions } from "./base";
 import { Buffer } from "buffer";
