+++ packages/wallets/src/evm/wallets/wallet-connect-v1.ts
@@ -2,7 +2,7 @@ import type { ConnectorData } from "../../lib/wagmi-core";
 import type { WalletConnectV1Connector } from "../connectors/wallet-connect-v1";
 import type WalletConnectProvider from "../connectors/wallet-connect-v1/walletconnect-legacy-provider";
 import { TWConnector, WagmiAdapter } from "../interfaces/tw-connector";
-import { walletIds } from "../walletIds";
+import { walletIds } from "../constants/walletIds";
 import { AbstractClientWallet, WalletOptions } from "./base";
 
 export type WalletConnectV1Options = {
