+++ packages/wallets/src/evm/wallets/wallet-connect.ts
@@ -4,7 +4,7 @@ import { TWConnector, WagmiAdapter } from "../interfaces/tw-connector";
 import { AbstractClientWallet, WalletOptions } from "./base";
 import type WalletConnectProvider from "@walletconnect/ethereum-provider";
 import { TW_WC_PROJECT_ID } from "../constants/wc";
-import { walletIds } from "../walletIds";
+import { walletIds } from "../constants/walletIds";
 
 export type WalletConnectOptions = {
   projectId?: string;
